+++ relic/scene_level_functionals.py
