--- som/namespace.py
@@ -1,2 +1,0 @@
-
-
