+++ vqa/safety_question_generation.py
@@ -6,17 +6,16 @@
 #
 import json
 import os
-import glob
 from vqa.scene_graph import TemporalGraph
 import random
-from vqa.dynamic_question_generation import find_episodes, extract_observations
+from relic.dynamic_question_generation import find_episodes, extract_observations
 from collections import defaultdict
 import numpy as np
 
-from vqa.scene_level_functionals import sample_keypoints, extrapolate_bounding_boxes, counterfactual_trajectory, \
+from relic.scene_level_functionals import sample_keypoints, extrapolate_bounding_boxes, counterfactual_trajectory, \
     counterfactual_stop, locate_crash_timestamp, move_around, predict_collision
 from vqa.object_node import box_overlap, transform_vec
-from vqa.dynamic_question_generation import extract_frames, generate_context_string
+from relic.dynamic_question_generation import extract_frames, generate_context_string
 
 
 def choices(question_type, graph):
